Ensure Monte-Carlo forcing dirs are run-specific
This is to avoid the possibility of concurrent runs that happen to be for the
same time period trying to create forcing files in the same directory.

With this change forcing files become truly ephemeral - used for 1 run only -
so they and their directories are removed at the end of the run.