Start MOHID run at spill hour; end it run_days later
Actually spill hour + 30 min to align with forcing fields,
but the 30 is hard coded in the Model.dat template file.

This implementation of spill hour makes results like beaching time relative to
the start of the run in all cases,
in contrast to what would happen if spill hour was implemented via
START_PARTIC_EMIT in Lagrangian.dat file.
It also avoids spending compute on time before the spill.